backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
motor==3.4.0
pydantic==2.7.4
pydantic-settings==2.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
openai==1.35.0
tensorflow-cpu==2.16.1
huggingface_hub==0.23.4
Pillow==10.3.0
numpy==1.26.4
firebase-admin==6.5.0
python-dotenv==1.0.1
email-validator==2.2.0
joblib==1.4.2
scikit-learn==1.5.1
google-generativeai==0.7.2
anthropic==0.34.2
